+++ src/trpc/routers/groups/delete.procedure.ts
@@ -1,14 +1,16 @@
 import { deleteGroup } from '@/lib/groups'
-import { baseProcedure } from '@/trpc/init'
+import { protectedProcedure } from '@/trpc/init'
+import { requireGroupOwner } from './authorization'
 import { z } from 'zod'
 
-export const deleteGroupProcedure = baseProcedure
+export const deleteGroupProcedure = protectedProcedure
   .input(
     z.object({
       groupId: z.string().min(1),
       participantId: z.string().optional(),
     }),
   )
-  .mutation(async ({ input: { groupId, participantId } }) => {
+  .mutation(async ({ ctx, input: { groupId, participantId } }) => {
+    await requireGroupOwner(ctx.auth.user.id, groupId)
     await deleteGroup(groupId, participantId)
   })
